Change the location of the test and the check for Pandas version
